pyproject.toml
[project]
name = "tune_jax"
version = "0.5.0"
authors = [{ name = "Robert Dyro" }]
dependencies = ["jax", "numpy", "tqdm", "protobuf", "tabulate"]
description = "Microbenchmarking hyperparameter tuning for JAX functions."
readme = "README.md"
license = { file = "LICENSE" }

[project.optional-dependencies]
test = ["absl-py", "pytest"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
preview = true
exclude = [
    ".git",
    "build",
    "__pycache__",
    "tune_jax/profile_reader/xplane_pb2.py",
]
line-length = 120
indent-width = 2
target-version = "py310"

[tool.ruff.lint]
ignore = [
    "C408",  # Unnecessary collection call
    "C901",  # Object names too complex
    "F841",  # Local variable is assigned to but never used
    "B904",  # Raise with from clause inside except block
    "E111",  # ignore indentation not being 4 (we use 2)
    "E114",  # ignore indentation not being 4 in comments
    "E731",  # allow local lambdas
]
select = ["B9", "C", "F", "W", "YTT", "ASYNC", "E"]
